Create index.html
october 20 2019

aded this to my blog website thing. this is so that I can begin to trnsition it into more of wht aaronsmusic dojo will be like. Eventually maybe it will become the everett music dojo if or when I get a brick nd mortar in everett of wherever i live