worked on canvas game
tried to work on canvas game but import paddle is not working. i am going to be a little wuss and play computer games now instead of continuing to work on this problem and solve it. i dont think i can import paddle with regular javascript syntax which is wierd because thats what i thought the guy was doing in the video. this will require much further research