made a little game
added a little div game to try and recreate the little red box that moves with the d pad on archive.html. It is linked within the canvas game which is linked from the main index.html. This was not succsesful so far as it was much more complicated than i rmemebr. the javascript is over 50 lines. I d not know why the javascript is note okring. I have not tried enteing a console.log in the .js to see if it is working, but i think the script tag is succesfully calling the .js because i am getting errors in the console referring to lines in that .js file. This is a big thing in my life right now, and I must go to bed before I suddenly collapse and die of not sleeping tomorrow ewhen i am at work trying to produce economic value. I have to go to bed right now at this very pomenet but instead I am typing vainy to see how many wpm I CAN GET AND HOW FEW ERRORS. THIS IS UTTERLY RIDICULOUS YEARS FROM NOW I WILL LOOK BACK AND SCOFF. RIGHT NOW I AM LOOKING BAC AT THE PREVIOUS SECONDS WHICH GE JUST PASSED AND I AM CURRENTLY SCOFFING BUT SILENTLY AND WITHOUT ANY EXPRESSIONFULNESS IN MY FACE. THIS IS UNEXPRESSIONED SCOFFING. Oops I had caps lock on that whole time. Now everyone wil lthink that IU was very angry when INfact I am merely sleepy.